+++ src/config.ts
@@ -4,60 +4,31 @@ import {
   IntegrationInstanceConfigFieldMap,
   IntegrationInstanceConfig,
 } from '@jupiterone/integration-sdk-core';
-import { Client } from './client';
+// import { Client } from './client';
 import { ParsedServiceAccountKeyFile } from './utils/parseServiceAccountKeyFile';
 
 export const instanceConfigFields: IntegrationInstanceConfigFieldMap = {
-  type: {
-    type: 'string',
-  },
-  project_id: {
-    type: 'string',
-  },
-  private_key_id: {
-    type: 'string',
-    mask: true,
-  },
-  private_key: {
-    type: 'string',
-    mask: true,
-  },
-  client_email: {
-    type: 'string',
-    mask: true,
-  },
-  client_id: {
-    type: 'string',
-  },
-  auth_uri: {
-    type: 'string',
-  },
-  token_uri: {
-    type: 'string',
-  },
-  auth_provider_x509_cert_url: {
-    type: 'string',
-  },
-  client_x509_cert_url: {
+  serviceAccountKeyFile: {
     type: 'string',
   },
 };
 
 export interface IntegrationConfig extends IntegrationInstanceConfig {
-  serviceAccountKeyConfig: ParsedServiceAccountKeyFile;
+  serviceAccountKeyFile: ParsedServiceAccountKeyFile;
 }
 
 export async function validateInvocation(
   context: IntegrationExecutionContext<IntegrationConfig>,
 ) {
   const { config } = context.instance;
+  console.log('config', config);
 
-  if (!config.serviceAccountKeyConfig) {
+  if (!config.serviceAccountKeyFile) {
     throw new IntegrationValidationError(
-      'Config requires all of {serviceAccountKeyConfig}',
+      'Config requires all of {serviceAccountKeyFile}',
     );
   }
 
-  const apiClient = new Client(config);
-  await apiClient.verifyAuthentication();
+  // const apiClient = new Client(config);
+  // await apiClient.verifyAuthentication();
 }
