+++ src/google-cloud/utils/createErrorProps.ts
