+++ src/google-cloud/utils/typeFunctions.ts
