+++ src/steps/users/index.ts
@@ -3,29 +3,26 @@ import {
   IntegrationStepExecutionContext,
 } from '@jupiterone/integration-sdk-core';
 
-import { Client } from '../../client';
 import { IntegrationConfig } from '../../config';
 import { Steps, Entities, Relationships } from '../constants';
+import { IdentityToolkitClient } from './client';
 import { createUserEntity, createProjectUserRelationship } from './converter';
 
 export async function fetchUsers({
   instance,
   jobState,
 }: IntegrationStepExecutionContext<IntegrationConfig>) {
-  const apiClient = new Client(instance.config);
+  const client = new IdentityToolkitClient(instance.config);
 
   await jobState.iterateEntities(
     { _type: Entities.PROJECT._type },
     async (projectEntity) => {
-      await apiClient.iterateUsers(
-        projectEntity.key as string,
-        async (user) => {
-          const userEntity = await jobState.addEntity(createUserEntity(user));
-          await jobState.addRelationship(
-            createProjectUserRelationship(projectEntity, userEntity),
-          );
-        },
-      );
+      await client.iterateUsers(projectEntity.key as string, async (user) => {
+        const userEntity = await jobState.addEntity(createUserEntity(user));
+        await jobState.addRelationship(
+          createProjectUserRelationship(projectEntity, userEntity),
+        );
+      });
     },
   );
 }
